Added 2 more unit tests. Refactored test class package
