+++ src/document-utils-api/src/test/java/ca/bc/gov/open/jag/documentutils/adobe/service/AemServiceImplTest.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.open.jag.documentutils.adobe; 
+package ca.bc.gov.open.jag.documentutils.adobe.service; 
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
